Moving spawnables tests into main suite
