+++ AutomatedTesting/Gem/PythonTests/Prefab/CMakeLists.txt
@@ -20,16 +20,16 @@ if(PAL_TRAIT_BUILD_TESTS_SUPPORTED AND PAL_TRAIT_BUILD_HOST_TOOLS)
         LABELS REQUIRES_tiaf
     )
 
-    ly_add_pytest(
-        NAME AutomatedTesting::PrefabTests_Periodic
-        TEST_SUITE periodic
-        TEST_SERIAL
-        PATH ${CMAKE_CURRENT_LIST_DIR}/TestSuite_Periodic.py
-        RUNTIME_DEPENDENCIES
-            Legacy::Editor
-            AZ::AssetProcessor
-            AutomatedTesting.Assets
-    )
+    #ly_add_pytest(
+    #    NAME AutomatedTesting::PrefabTests_Periodic
+    #    TEST_SUITE periodic
+    #    TEST_SERIAL
+    #    PATH ${CMAKE_CURRENT_LIST_DIR}/TestSuite_Periodic.py
+    #    RUNTIME_DEPENDENCIES
+    #        Legacy::Editor
+    #        AZ::AssetProcessor
+    #        AutomatedTesting.Assets
+    #)
 
     ly_add_pytest(
         NAME AutomatedTesting::PrefabLauncherTests_Periodic
