+++ AutomatedTesting/Gem/PythonTests/Prefab/TestSuite_Periodic.py
@@ -7,7 +7,7 @@ SPDX-License-Identifier: Apache-2.0 OR MIT
 
 import pytest
 
-from ly_test_tools.o3de.editor_test import EditorBatchedTest, EditorTestSuite
+from ly_test_tools.o3de.editor_test import EditorBatchedTest, EditorSingleTest, EditorTestSuite
 
 
 @pytest.mark.SUITE_periodic
@@ -18,34 +18,6 @@ class TestAutomationNoOverrides(EditorTestSuite):
     # These tests will execute with prefab outliner overrides disabled
     EditorTestSuite.global_extra_cmdline_args.append("--regset=O3DE/Preferences/Prefabs/EnableOutlinerOverrideManagement=false")
 
-    # Spawnables Tests
-
-    class test_SC_Spawnables_DespawnOnEntityDeactivate(EditorBatchedTest):
-        from .tests.spawnables import SC_Spawnables_DespawnOnEntityDeactivate as test_module
-
-    class test_SC_Spawnables_EntityClearedOnGameModeExit(EditorBatchedTest):
-        from .tests.spawnables import SC_Spawnables_EntityClearedOnGameModeExit as test_module
-
-    class test_SC_Spawnables_MultipleSpawnsFromSingleTicket(EditorBatchedTest):
-        from .tests.spawnables import SC_Spawnables_MultipleSpawnsFromSingleTicket as test_module
-
-    class test_SC_Spawnables_NestedSpawn(EditorBatchedTest):
-        from .tests.spawnables import SC_Spawnables_NestedSpawn as test_module
-
-    class test_SC_Spawnables_SimpleSpawnAndDespawn(EditorBatchedTest):
-        from .tests.spawnables import SC_Spawnables_SimpleSpawnAndDespawn as test_module
-
-    class test_Lua_Spawnables_DespawnOnEntityDeactivate(EditorBatchedTest):
-        from .tests.spawnables import Lua_Spawnables_DespawnOnEntityDeactivate as test_module
-
-    class test_Lua_Spawnables_EntityClearedOnGameModeExit(EditorBatchedTest):
-        from .tests.spawnables import Lua_Spawnables_EntityClearedOnGameModeExit as test_module
-
-    class test_Lua_Spawnables_MultipleSpawnsFromSingleTicket(EditorBatchedTest):
-        from .tests.spawnables import Lua_Spawnables_MultipleSpawnsFromSingleTicket as test_module
-
-    class test_Lua_Spawnables_NestedSpawn(EditorBatchedTest):
-        from .tests.spawnables import Lua_Spawnables_NestedSpawn as test_module
-
-    class test_Lua_Spawnables_SimpleSpawnAndDespawn(EditorBatchedTest):
-        from .tests.spawnables import Lua_Spawnables_SimpleSpawnAndDespawn as test_module
+    @pytest.mark.skip(reason="Single test case to avoid suite failure. Can be removed when other tests are added.")
+    class test_DummyTest(EditorSingleTest):
+        pass
